修改工作机制变成时长固定，总量和工作速度改变，抽象出一个工作类

+++ src/main/java/com/science/game/cache/config/ItemConfigCache.java
@@ -24,7 +24,7 @@ public class ItemConfigCache implements IConfigCache {
 		config.setDevelopPoint(getInt(values, 4));
 		config.setEffect(getInt(values, 5));
 		config.setAge(getInt(values, 6));
-		config.setUnitTotal(getLong(values, 7));
+		config.setUnitTotal(getInt(values, 7));
 		config.setUnitVelocity(getInt(values, 8));
 
 		itemMap.put(config.getItemId(), config);
