+++ src/main/java/com/science/game/entity/config/JobConfig.java
@@ -7,7 +7,7 @@ public class JobConfig {
 	private int id;
 	private String job;
 	// 单位生产时长
-	private long unitTotal;
+	private int unitTotal;
 	// 单位产量
 	private int unitVelocity;
 }
